Tests/LibWeb/Text/expected/show-non-modal-dialog.txt

dialog.open state before show(): false
dialog.open state after show(): true
